start event handler cleanup
