+++ src/data/java/com/github/minecraftschurlimods/arsmagicalegacy/data/AMDatagen.java
@@ -1,7 +1,7 @@
 package com.github.minecraftschurlimods.arsmagicalegacy.data;
 
 import com.github.minecraftschurlimods.arsmagicalegacy.api.ArsMagicaAPI;
-import com.github.minecraftschurlimods.arsmagicalegacy.common.EventHandler;
+import com.github.minecraftschurlimods.arsmagicalegacy.common.handler.EventHandler;
 import net.minecraft.data.DataGenerator;
 import net.minecraftforge.common.data.ExistingFileHelper;
 import net.minecraftforge.common.data.LanguageProvider;
